feat(web): add tooltip and toast providers
