--- apps/web/components/providers.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import * as React from 'react';
-import { ThemeProvider as NextThemesProvider } from 'next-themes';
-
-export function Providers({ children }: { children: React.ReactNode }) {
-	return (
-		<NextThemesProvider
-			attribute="class"
-			defaultTheme="system"
-			enableSystem
-			disableTransitionOnChange
-			enableColorScheme
-		>
-			{children}
-		</NextThemesProvider>
-	);
-}
