+++ apps/web/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./*"],
+      "@/*": ["./src/*"],
       "@rp/ui/*": ["../../packages/ui/src/*"]
     },
     "plugins": [
